kos-py/pykos/requirements.txt
# requirements.txt

grpcio
protobuf>=5.29.0
googleapis-common-protos
nest-asyncio
